+++ src/viksa_ai/__init__.py
@@ -2,7 +2,7 @@
 
 from typing import TYPE_CHECKING
 
-__version__ = "0.2.1"
+__version__ = "0.2.4"
 
 __all__ = [
     "ViksaApiError",